pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "carmel"
dynamic = ["version"]
description = "Closed-loop campaign manager for predictive chemical kinetics"
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.1",
    "pandas>=2.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
]

[project.scripts]
carmel = "Carmel:cli"

[tool.setuptools.dynamic]
version = {attr = "carmel.version.__version__"}

[tool.setuptools.packages.find]
include = ["carmel*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["."]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 90

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
